Refactor: Create/rename some packages for better overview
