+++ frontend/src/components/spot-list/SpotList.tsx
