--- frontend/src/components/ui/CustomListItemTextInput.tsx
@@ -1,30 +1,0 @@
-import {FormControl, ListItem, TextField} from "@mui/material";
-import {ChangeEvent} from "react";
-
-type CustomListItemProps = {
-    name: string
-    label: string
-    onChange(event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>): void;
-    value?: string
-    editable: boolean
-}
-export default function CustomListItemTextInput(props: CustomListItemProps) {
-
-
-    function onChange(event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) {
-        props.onChange(event)
-    }
-
-    return (
-        <ListItem>
-            <FormControl fullWidth={true} margin={"dense"}>
-                <TextField name={props.name}
-                           type={"text"}
-                           label={props.label}
-                           value={props.value}
-                           disabled={!props.editable}
-                           onChange={onChange}/>
-            </FormControl>
-        </ListItem>
-    )
-}
