+++ frontend/src/components/ui/custom-mui-components/CustomAccordion.tsx
