+++ frontend/src/components/ui/custom-mui-components/CustomCard.tsx
